backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-dateutil==2.8.2
cryptography==41.0.7
